offers+postulation

+++ database/migrations/2025_03_29_212739_create_postulation_table.php
@@ -7,22 +7,23 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up()
     {
-        Schema::create('postulation', function (Blueprint $table) {
-            $table->bigIncrements('id_postulation');
-            $table->unsignedBigInteger('id_users'); 
-            $table->unsignedBigInteger('id_offers'); 
-            $table->string('cv',255); 
-            $table->string('motivation_letter',255); 
-            $table->string('status',50)->default('Pending'); 
-            $table->timestamps(); 
-            // Définition des clés étrangères
-            $table->foreign('id_users')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('id_offers')->references('id_offers')->on('offers')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('postulations')) {
+            Schema::create('postulations', function (Blueprint $table) {
+                $table->bigIncrements('id_postulations');
+                $table->unsignedBigInteger('id_users'); 
+                $table->unsignedBigInteger('id_offers'); 
+                $table->string('cv',255); 
+                $table->string('motivation_letter',255); 
+                $table->string('status',50)->default('Pending'); 
+                $table->timestamps(); 
+                // Définition des clés étrangères
+                $table->foreign('id_users')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('id_offers')->references('id_offers')->on('offers')->onDelete('cascade');
+            });
+        }
     }
-
     public function down()
     {
-        Schema::dropIfExists('postulation');
+        Schema::dropIfExists('postulations');
     }
 };
